+++ src/gql/interfaces/SeomaticInterface.php
@@ -11,8 +11,14 @@
 
 namespace nystudio107\seomatic\gql\interfaces;
 
+use nystudio107\seomatic\gql\arguments\FrontendContainerArguments;
+use nystudio107\seomatic\gql\arguments\SitemapArguments;
+use nystudio107\seomatic\gql\arguments\SitemapIndexArguments;
+use nystudio107\seomatic\gql\resolvers\FrontendContainerResolver;
+use nystudio107\seomatic\gql\resolvers\SitemapResolver;
 use nystudio107\seomatic\gql\types\generators\SeomaticGenerator;
 
+use nystudio107\seomatic\gql\types\FileContentsType;
 use nystudio107\seomatic\models\FrontendTemplateContainer;
 use nystudio107\seomatic\models\MetaJsonLdContainer;
 use nystudio107\seomatic\models\MetaLinkContainer;
@@ -47,6 +53,10 @@ class SeomaticInterface extends BaseInterfaceType
         'metaJsonLdContainer' => MetaJsonLdContainer::CONTAINER_TYPE,
         'metaSiteVarsContainer' => MetaSiteVars::CONTAINER_TYPE,
         'frontendTemplateContainer' => FrontendTemplateContainer::CONTAINER_TYPE,
+    ];
+
+    const DEPRECATED_GRAPH_QL_FIELDS = [
+        'frontendTemplateContainer' => 'This query is deprecated and will be removed in the future. You should use `frontendTemplates` instead.',
     ];
 
     /**
@@ -94,12 +104,43 @@ class SeomaticInterface extends BaseInterfaceType
     {
         $fields = [];
         foreach (self::GRAPH_QL_FIELDS as $key => $value) {
+
             $fields[$key] = [
                 'name' => $key,
                 'type' => Type::string(),
                 'description' => 'The '.$value.' SEOmatic container.',
             ];
+            if (isset(self::DEPRECATED_GRAPH_QL_FIELDS[$key])) {
+                $fields[$key]['deprecationReason'] = self::DEPRECATED_GRAPH_QL_FIELDS[$key];
+            }
         }
+
+        $fields['sitemaps'] = [
+            'name' => 'sitemaps',
+            'args' => SitemapArguments::getArguments(),
+            'type' => Type::listOf(FileContentsType::getType()),
+            'resolve' => SitemapResolver::class .'::getSitemaps'
+        ];
+
+        $fields['sitemapIndexes'] = [
+            'name' => 'sitemapIndexes',
+            'args' => SitemapIndexArguments::getArguments(),
+            'type' => Type::listOf(FileContentsType::getType()),
+            'resolve' => SitemapResolver::class .'::getSitemapIndexes'
+        ];
+
+        $fields['sitemapStyles'] = [
+            'name' => 'sitemapStyles',
+            'type' => FileContentsType::getType(),
+            'resolve' => SitemapResolver::class .'::getSitemapStyles'
+        ];
+
+        $fields['frontendTemplates'] = [
+            'name' => 'frontendTemplates',
+            'args' => FrontendContainerArguments::getArguments(),
+            'type' => Type::listOf(FileContentsType::getType()),
+            'resolve' => FrontendContainerResolver::class .'::getContainerFiles'
+        ];
 
         return $fields;
     }