+++ src/helpers/Text.php
@@ -19,6 +19,7 @@ use craft\elements\db\MatrixBlockQuery;
 use craft\elements\db\TagQuery;
 use craft\elements\MatrixBlock;
 use craft\elements\Tag;
+use craft\helpers\HtmlPurifier;
 
 use yii\base\InvalidConfigException;
 
@@ -71,7 +72,8 @@ class Text
         $result = $string;
 
         if (!empty($string)) {
-            $string = strip_tags($string);
+            $string = HtmlPurifier::process($string, ['HTML.Allowed' => '']);
+            $string = html_entity_decode($string, ENT_NOQUOTES, 'UTF-8');
             $result = (string)Stringy::create($string)->truncate($length, $substring);
         }
 
@@ -95,7 +97,8 @@ class Text
         $result = $string;
 
         if (!empty($string)) {
-            $string = strip_tags($string);
+            $string = HtmlPurifier::process($string, ['HTML.Allowed' => '']);
+            $string = html_entity_decode($string, ENT_NOQUOTES, 'UTF-8');
             $result = (string)Stringy::create($string)->safeTruncate($length, $substring);
         }
 
@@ -395,7 +398,8 @@ class Text
         $str = rawurldecode($str);
         // Remove any linebreaks
         $str = (string)preg_replace("/\r|\n/", "", $str);
-        $str = strip_tags($str);
+        $str = HtmlPurifier::process($str, ['HTML.Allowed' => '']);
+        $str = html_entity_decode($str, ENT_NOQUOTES, 'UTF-8');
         // Remove any embedded Twig code
         $str = preg_replace('/{{.*?}}/', '', $str);
         $str = preg_replace('/{%.*?%}/', '', $str);
@@ -418,7 +422,8 @@ class Text
     public static function smartStripTags($str)
     {
         $str = str_replace('<', ' <', $str);
-        $str = strip_tags($str);
+        $str = HtmlPurifier::process($str, ['HTML.Allowed' => '']);
+        $str = html_entity_decode($str, ENT_NOQUOTES, 'UTF-8');
         $str = str_replace('  ', ' ', $str);
 
         return $str;
@@ -445,7 +450,8 @@ class Text
             $text = mb_convert_encoding($text, 'UTF-8', 'UTF-8');
         }
         // Strip HTML tags
-        $text = strip_tags($text);
+        $text = HtmlPurifier::process($text, ['HTML.Allowed' => '']);
+        $text = html_entity_decode($text, ENT_NOQUOTES, 'UTF-8');
         // Remove excess whitespace
         $text = preg_replace('/\s{2,}/u', ' ', $text);
         // Decode any HTML entities