+++ src/models/metalink/CanonicalLink.php
@@ -78,7 +78,7 @@ class CanonicalLink extends MetaLink
             // Don't render a canonical URL if this page isn't meant to be indexed as per:
             // https://www.reddit.com/r/TechSEO/comments/8yahdr/2_questions_about_the_canonical_tag/e2dey9i/?context=1
             $robots = Seomatic::$seomaticVariable->tag->get('robots');
-            if ($robots !== null && $robots->include && !Seomatic::$previewingMetaContainers) {
+            if ($robots !== null && $robots->include && !Seomatic::$previewingMetaContainers && !Seomatic::$settings->alwaysIncludeCanonicalUrls) {
                 $robotsArray = $robots->renderAttributes();
                 $content = $robotsArray['content'] ?? '';
                 if (!empty($content)) {